+++ Sources/WalletConnectRelay/ClientAuth/JWT/JWT.swift
@@ -28,31 +28,3 @@ struct JWT: Codable, Equatable {
         return [headerString, claimsString, signature].joined(separator: ".")
     }
 }
-
-extension JWT {
-    struct Header: Codable, Equatable {
-        var alg: String!
-        let typ: String = "JWT"
-
-        func encode() throws -> String  {
-            let jsonEncoder = JSONEncoder()
-            jsonEncoder.dateEncodingStrategy = .secondsSince1970
-            let data = try jsonEncoder.encode(self)
-            return JWTEncoder.base64urlEncodedString(data: data)
-        }
-    }
-}
-
-extension JWT {
-    struct Claims: Codable, Equatable {
-        let iss: String
-        let sub: String
-        func encode() throws -> String  {
-            let jsonEncoder = JSONEncoder()
-            jsonEncoder.dateEncodingStrategy = .secondsSince1970
-            let data = try jsonEncoder.encode(self)
-            return JWTEncoder.base64urlEncodedString(data: data)
-        }
-    }
-}
-
